Android: include feature flags in GitHub issue and email reports
Flag context (autoCopy, otpRelay, hideSyncedClipboard,
autoClearSyncedClipboard, imageSync) and paired-Mac count were only
included in shared logs, not in the pre-filled GitHub issue or support
email bodies. Move the flag collection from LogShareExporter into
SupportLinks.diagnosticsContext so all three reporting paths share it,
and add the previously missing otpRelay flag.

Motivated by #109, where the report lacked the settings needed to
narrow down the cause.
